Initial Commit
This Project was originally forked off the
processing/processing-library-template repository; however, this was
limiting and therefore I have made a new repository for the foreseeable
future.

So lets talk a little about what this project is. This Project is a
Library that has been created for Processing. Processing is a PDE and I
will let the website and your curiosity explain it in detail.

This Library is designed with ease of use in mind. I wanted to make
handling keys in processing a little more advanced but also keep
simplicity in mind. In the midst of chaos going around in my brain I
came up with this.

NOTICE: I want to do some renaming of stuff like the class name and
maybe even shorting of the functions in the class. Any suggestions are
greatly appreciated.

During development I attempted to leave the current key handling in tact
and was successful; however, I did want to be able to add extra into it
but without full success. The issue lies in the fact of how processing
executes the keyEvent vs keyPressed/keyReleased/keyTyped methods. First
it executes the keyPressed/keyReleased/keyTyped and resolves, then it
executes keyTyped. Due to this I cannot make it work; however, I do know
what to do to fix it now and this will be dealt with at a later date.

Another issue that came up was when the Window loses and gains focus
there is no current way to determine this. I attempted to use the built
in controls but to no avail was I successful. This will be looked into
more at a later time.

As of right now all the java-doc in the Library should be able to
explain everything about the Library as it currently stands. I will be
double checking this in the near future.

This is a functioning version of the Library; however, it needs a lot of
work done to it to make everything correct, look better, and perform
better.

Library SHOULD be thread-safe; however, this has not been strain tested
and confirmed. A lot more testing needs to be done also.